D1: 拆 doctor_exam_question_rule_edit_dialog.dart 的 3 个 Dialog
ChoiceSettingDialog / ItemSlotEditDialog / QuestionScoreConditionEditDialog
3 个互相独立的 Dialog 各自抽到独立文件，原 955 行的 doctor_exam_question_rule_edit_dialog.dart 删除，
doctor_exam_question_rule_edit.dart 改为直接 import 3 个新文件。

零行为变更，flutter analyze 零新增告警。
