+++ frontend/lib/widgets/ui/doctor/doctor_exam_question_rule_edit.dart
@@ -11,7 +11,9 @@ import 'package:aphasia_recovery/widgets/ui/common/common.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:aphasia_recovery/mixin/eval_rule_mixin.dart';
-import 'doctor_exam_question_rule_edit_dialog.dart';
+import 'doctor_choice_setting_dialog.dart';
+import 'doctor_item_slot_edit_dialog.dart';
+import 'doctor_question_score_condition_edit_dialog.dart';
 import '../../../models/question/question.dart';
 import '../../../models/rules.dart';
 import '../../../utils/common_widget_function.dart';
